requirements.txt
MouseInfo==0.1.3
numpy==1.21.6
opencv-python==4.6.0.66
Pillow==9.2.0
psutil==5.9.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
